# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: ProxySieve local control-plane API
  version: 0.1.0
  description: >-
    The authenticated local Admin API exposed by ProxySieve. This contract
    describes the implemented surface; planned resources are intentionally omitted.
servers:
  - url: http://127.0.0.1:9090
    description: Default local-only listener
...
paths:
...
  /api/v1/policies:
    get:
      tags: [Policies]
      parameters: [{$ref: '#/components/parameters/Limit'}, {name: after, in: query, schema: {type: string}}]
      responses:
        '200': {description: Revisioned policy inventory, content: {application/json: {schema: {$ref: '#/components/schemas/PolicyPage'}}}}
    post:
      tags: [Policies]
      security: [sessionCookie: [], csrfToken: []]
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, required: [policy], properties: {policy: {$ref: '#/components/schemas/Policy'}}}}}
      responses:
        '201': {description: Policy inventory record created, content: {application/json: {schema: {$ref: '#/components/schemas/PolicyResponse'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
  /api/v1/policies/{policyId}:
    parameters: [{name: policyId, in: path, required: true, schema: {type: string}}]
    get:
      tags: [Policies]
      responses: {'200': {description: Policy record, content: {application/json: {schema: {$ref: '#/components/schemas/PolicyResponse'}}}}, '404': {description: Policy not found}}
    patch:
      tags: [Policies]
      security: [sessionCookie: [], csrfToken: []]
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [policy, revision], properties: {policy: {$ref: '#/components/schemas/Policy'}, revision: {type: integer, minimum: 1}}}}}}
      responses: {'200': {description: Updated policy record, content: {application/json: {schema: {$ref: '#/components/schemas/PolicyResponse'}}}}, '400': {$ref: '#/components/responses/BadRequest'}, '409': {description: Stale revision}}
    delete:
      tags: [Policies]
      security: [sessionCookie: [], csrfToken: []]
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [revision], properties: {revision: {type: integer, minimum: 1}}}}}}
      responses: {'204': {description: Policy deleted}, '409': {description: Stale revision}}
  /api/v1/policies/{policyId}/simulate:
    parameters: [{name: policyId, in: path, required: true, schema: {type: string}}]
    post:
      tags: [Policies]
      security: [sessionCookie: [], csrfToken: []]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SimulationRequest'}}}}
      responses: {'200': {description: Simulation result; does not activate runtime routing, content: {application/json: {schema: {$ref: '#/components/schemas/SimulationResponse'}}}}, '400': {$ref: '#/components/responses/BadRequest'}}
...
components:
  securitySchemes:
    sessionCookie: {type: apiKey, in: cookie, name: proxysieve_session}
    csrfToken: {type: apiKey, in: header, name: X-CSRF-Token}
...
  parameters:
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 1000, default: 100}}
...
  responses:
    BadRequest: {description: Request was not accepted, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: {type: object, required: [code, message], properties: {code: {type: string}, message: {type: string}}}
...
    Policy:
      type: object
      required: [version, id, name, rules]
      properties:
        version: {type: integer, const: 1}
        id: {type: string}
        name: {type: string, minLength: 1, maxLength: 256}
        rules: {type: array, maxItems: 10000, items: {$ref: '#/components/schemas/PolicyRule'}}
    PolicyRule:
      type: object
      required: [id, name, priority, enabled, stop_processing, conditions, actions]
      properties:
        id: {type: string}
        name: {type: string}
        priority: {type: integer}
        enabled: {type: boolean}
        stop_processing: {type: boolean}
        conditions: {type: object}
        actions: {type: array, maxItems: 16, items: {type: object}}
    PolicyRecord:
      type: object
      required: [policy, revision]
      properties: {policy: {$ref: '#/components/schemas/Policy'}, revision: {type: integer, minimum: 1}}
    PolicyPage:
      type: object
      required: [items, next_after]
      properties: {items: {type: array, items: {$ref: '#/components/schemas/PolicyRecord'}}, next_after: {type: string}}
    PolicyResponse:
      type: object
      required: [policy, runtime_active, activation, runtime_revision]
      properties: {policy: {$ref: '#/components/schemas/PolicyRecord'}, runtime_active: {type: boolean}, activation: {type: string, enum: [active, staged]}, runtime_revision: {type: integer, minimum: 0}}
...
    SimulationRequest:
      type: object
      required: [listener, protocol, host, port]
      properties: {listener: {type: string}, client_id: {type: string}, protocol: {type: string}, scheme: {type: string}, host: {type: string}, port: {type: integer, minimum: 1, maximum: 65535}, method: {type: string}, path: {type: string}, destination_ip: {type: string}, timestamp: {type: string, format: date-time}}
    SimulationResponse:
      type: object
      required: [policy_id, revision, outcome, actions, matched_rule_ids, rules, unknown_fields, simulation_only, runtime_active]
      properties: {policy_id: {type: string}, revision: {type: integer}, outcome: {type: string}, actions: {type: array, items: {type: object}}, matched_rule_ids: {type: array, items: {type: string}}, rules: {type: array, items: {type: object}}, unknown_fields: {type: array, items: {type: string}}, simulation_only: {type: boolean, const: true}, runtime_active: {type: boolean}}
...